+++ keystorage/src/main/java/tech/pegasys/web3signer/keystorage/hashicorp/config/ConnectionParameters.java
@@ -12,39 +12,40 @@
  */
 package tech.pegasys.web3signer.keystorage.hashicorp.config;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 import java.net.URI;
+import java.net.http.HttpClient;
 import java.util.Optional;
 
 public class ConnectionParameters {
   private static final Long DEFAULT_TIMEOUT_MILLISECONDS = 10_000L;
   private static final Integer DEFAULT_SERVER_PORT = 8200;
-  private final String serverHost;
-  private final int serverPort;
   private final Optional<TlsOptions> tlsOptions;
   private final long timeoutMs;
+  private final HttpClient.Version httpProtocolVersion;
 
   private final URI vaultURI;
 
+  public static Builder newBuilder() {
+    return new Builder();
+  }
+
   /* Optional parameters will be set to their defaults when connecting */
-  public ConnectionParameters(
+  private ConnectionParameters(
       final String serverHost,
       final Optional<Integer> serverPort,
       final Optional<TlsOptions> tlsOptions,
-      final Optional<Long> timeoutMs) {
-    this.serverHost = serverHost;
-    this.serverPort = serverPort.orElse(DEFAULT_SERVER_PORT);
+      final Optional<Long> timeoutMs,
+      final Optional<HttpClient.Version> httpProtocolVersion) {
     this.tlsOptions = tlsOptions;
     this.timeoutMs = timeoutMs.orElse(DEFAULT_TIMEOUT_MILLISECONDS);
+    this.httpProtocolVersion = httpProtocolVersion.orElse(HttpClient.Version.HTTP_2);
     final String scheme = tlsOptions.isPresent() ? "https" : "http";
-    this.vaultURI = URI.create(String.format("%s://%s:%d", scheme, serverHost, this.serverPort));
-  }
-
-  public String getServerHost() {
-    return serverHost;
-  }
-
-  public int getServerPort() {
-    return serverPort;
+    this.vaultURI =
+        URI.create(
+            String.format(
+                "%s://%s:%d", scheme, serverHost, serverPort.orElse(DEFAULT_SERVER_PORT)));
   }
 
   public Optional<TlsOptions> getTlsOptions() {
@@ -57,5 +58,50 @@ public class ConnectionParameters {
 
   public URI getVaultURI() {
     return vaultURI;
+  }
+
+  public HttpClient.Version getHttpProtocolVersion() {
+    return httpProtocolVersion;
+  }
+
+  public static final class Builder {
+    private String serverHost;
+    private Optional<Integer> serverPort = Optional.empty();
+    private Optional<TlsOptions> tlsOptions = Optional.empty();
+    private Optional<Long> timeoutMs = Optional.empty();
+    private Optional<HttpClient.Version> httpProtocolVersion = Optional.empty();
+
+    Builder() {}
+
+    public Builder withServerHost(final String serverHost) {
+      this.serverHost = serverHost;
+      return this;
+    }
+
+    public Builder withServerPort(final Integer serverPort) {
+      this.serverPort = Optional.ofNullable(serverPort);
+      return this;
+    }
+
+    public Builder withTlsOptions(final TlsOptions tlsOptions) {
+      this.tlsOptions = Optional.ofNullable(tlsOptions);
+      return this;
+    }
+
+    public Builder withTimeoutMs(final Long timeoutMs) {
+      this.timeoutMs = Optional.ofNullable(timeoutMs);
+      return this;
+    }
+
+    public Builder withHttpProtocolVersion(final HttpClient.Version httpProtocolVersion) {
+      this.httpProtocolVersion = Optional.ofNullable(httpProtocolVersion);
+      return this;
+    }
+
+    public ConnectionParameters build() {
+      checkNotNull(serverHost, "Hashicorp host cannot be null");
+      return new ConnectionParameters(
+          serverHost, serverPort, tlsOptions, timeoutMs, httpProtocolVersion);
+    }
   }
 }